chore: get dark mode system default
