--- app/components/btnDarkMode/DarkMode.tsx
@@ -1,29 +1,0 @@
-'use client'
-
-import React, { useState, useEffect } from 'react'
-import { useTheme } from 'next-themes'
-import { FaMoon, FaSun } from 'react-icons/fa'
-import styles from './DarkMode.module.scss'
-
-const DarkModeButton = () => {
-  const [mounted, setMounted] = useState(false)
-  const { theme, setTheme } = useTheme()
-
-  useEffect(() => {
-    setMounted(true)
-  }, [])
-
-  if (!mounted) {
-    return null
-  }
-
-  return (
-    <button
-      className={`${styles.btn_root}`}
-      onClick={e => theme === 'dark' ? setTheme('light') : setTheme('dark')}>
-        {theme === 'dark' ? <FaMoon /> : <FaSun />}
-    </button>
-  )
-}
-
-export default DarkModeButton
